LPD-35630 SF - Rename classes to match base class
