+++ modules/apps/adaptive-media/adaptive-media-editor-configuration/src/test/java/com/liferay/adaptive/media/editor/configuration/internal/BlogsAMEditorConfigContributorTest.java
@@ -6,28 +6,30 @@
 package com.liferay.adaptive.media.editor.configuration.internal;
 
 import com.liferay.adaptive.media.image.item.selector.AMImageFileEntryItemSelectorReturnType;
+import com.liferay.blogs.item.selector.criterion.BlogsItemSelectorCriterion;
 import com.liferay.item.selector.ItemSelector;
 import com.liferay.item.selector.ItemSelectorCriterion;
 import com.liferay.item.selector.ItemSelectorReturnType;
 import com.liferay.item.selector.criteria.FileEntryItemSelectorReturnType;
 import com.liferay.item.selector.criteria.audio.criterion.AudioItemSelectorCriterion;
 import com.liferay.item.selector.criteria.file.criterion.FileItemSelectorCriterion;
-import com.liferay.journal.item.selector.criterion.JournalItemSelectorCriterion;
+import com.liferay.item.selector.criteria.image.criterion.ImageItemSelectorCriterion;
+import com.liferay.item.selector.criteria.upload.criterion.UploadItemSelectorCriterion;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.json.JSONUtil;
 import com.liferay.portal.kernel.portlet.RequestBackedPortletURLFactory;
 import com.liferay.portal.kernel.test.ReflectionTestUtil;
+import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.test.rule.LiferayUnitTestRule;
 
-import java.util.Collection;
-import java.util.Collections;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.function.Predicate;
 
 import javax.portlet.PortletURL;
 
@@ -44,7 +46,7 @@ import org.skyscreamer.jsonassert.JSONAssert;
 /**
  * @author Sergio González
  */
-public class AMJournalEditorConfigContributorTest {
+public class BlogsAMEditorConfigContributorTest {
 
 	@ClassRule
 	@Rule
@@ -85,25 +87,25 @@ public class AMJournalEditorConfigContributorTest {
 
 		Mockito.when(
 			_itemSelector.getItemSelectorCriteria(
-				"journalItemSelectorCriterionFileEntryItemSelectorReturnType")
+				"blogsItemSelectorCriterionFileEntryItemSelectorReturnType")
 		).thenReturn(
 			_getBlogsItemSelectorCriterionFileEntryItemSelectorReturnType()
 		);
 
 		JSONObject originalJSONObject = JSONUtil.put(
 			"filebrowserImageBrowseLinkUrl",
-			"journalItemSelectorCriterionFileEntryItemSelectorReturnType");
+			"blogsItemSelectorCriterionFileEntryItemSelectorReturnType");
 
 		JSONObject jsonObject = JSONFactoryUtil.createJSONObject(
 			originalJSONObject.toString());
 
-		AMJournalEditorConfigContributor amBlogsEditorConfigContributor =
-			new AMJournalEditorConfigContributor();
+		BlogsAMEditorConfigContributor blogsAMEditorConfigContributor =
+			new BlogsAMEditorConfigContributor();
 
 		ReflectionTestUtil.setFieldValue(
-			amBlogsEditorConfigContributor, "_itemSelector", _itemSelector);
+			blogsAMEditorConfigContributor, "_itemSelector", _itemSelector);
 
-		amBlogsEditorConfigContributor.populateConfigJSONObject(
+		blogsAMEditorConfigContributor.populateConfigJSONObject(
 			jsonObject, _inputEditorTaglibAttributes, _themeDisplay,
 			_requestBackedPortletURLFactory);
 
@@ -138,25 +140,25 @@ public class AMJournalEditorConfigContributorTest {
 
 		Mockito.when(
 			_itemSelector.getItemSelectorCriteria(
-				"journalItemSelectorCriterionFileEntryItemSelectorReturnType")
+				"blogsItemSelectorCriterionFileEntryItemSelectorReturnType")
 		).thenReturn(
 			_getBlogsItemSelectorCriterionFileEntryItemSelectorReturnType()
 		);
 
 		JSONObject originalJSONObject = JSONUtil.put(
 			"filebrowserImageBrowseLinkUrl",
-			"journalItemSelectorCriterionFileEntryItemSelectorReturnType");
+			"blogsItemSelectorCriterionFileEntryItemSelectorReturnType");
 
 		JSONObject jsonObject = JSONFactoryUtil.createJSONObject(
 			originalJSONObject.toString());
 
-		AMJournalEditorConfigContributor amJournalEditorConfigContributor =
-			new AMJournalEditorConfigContributor();
+		BlogsAMEditorConfigContributor blogsAMEditorConfigContributor =
+			new BlogsAMEditorConfigContributor();
 
 		ReflectionTestUtil.setFieldValue(
-			amJournalEditorConfigContributor, "_itemSelector", _itemSelector);
+			blogsAMEditorConfigContributor, "_itemSelector", _itemSelector);
 
-		amJournalEditorConfigContributor.populateConfigJSONObject(
+		blogsAMEditorConfigContributor.populateConfigJSONObject(
 			jsonObject, _inputEditorTaglibAttributes, _themeDisplay,
 			_requestBackedPortletURLFactory);
 
@@ -190,7 +192,7 @@ public class AMJournalEditorConfigContributorTest {
 
 		Mockito.when(
 			_itemSelector.getItemSelectorCriteria(
-				"journalItemSelectorCriterionFileEntryItemSelectorReturnType")
+				"blogsItemSelectorCriterionFileEntryItemSelectorReturnType")
 		).thenReturn(
 			_getBlogsItemSelectorCriterionFileEntryItemSelectorReturnType()
 		);
@@ -200,16 +202,16 @@ public class AMJournalEditorConfigContributorTest {
 				"extraPlugins", "ae_placeholder,ae_selectionregion,ae_uicore"
 			).put(
 				"filebrowserImageBrowseLinkUrl",
-				"journalItemSelectorCriterionFileEntryItemSelectorReturnType"
+				"blogsItemSelectorCriterionFileEntryItemSelectorReturnType"
 			).toString());
 
-		AMJournalEditorConfigContributor amJournalEditorConfigContributor =
-			new AMJournalEditorConfigContributor();
+		BlogsAMEditorConfigContributor blogsAMEditorConfigContributor =
+			new BlogsAMEditorConfigContributor();
 
 		ReflectionTestUtil.setFieldValue(
-			amJournalEditorConfigContributor, "_itemSelector", _itemSelector);
+			blogsAMEditorConfigContributor, "_itemSelector", _itemSelector);
 
-		amJournalEditorConfigContributor.populateConfigJSONObject(
+		blogsAMEditorConfigContributor.populateConfigJSONObject(
 			jsonObject, _inputEditorTaglibAttributes, _themeDisplay,
 			_requestBackedPortletURLFactory);
 
@@ -219,37 +221,97 @@ public class AMJournalEditorConfigContributorTest {
 	}
 
 	@Test
-	public void testAddAMImageFileEntryItemSelectorReturnType() {
-		AMJournalEditorConfigContributor amJournalEditorConfigContributor =
-			new AMJournalEditorConfigContributor();
+	public void testAddAMImageFileEntryItemSelectorReturnType()
+		throws Exception {
 
-		JournalItemSelectorCriterion journalItemSelectorCriterion =
-			new JournalItemSelectorCriterion();
+		BlogsAMEditorConfigContributor blogsAMEditorConfigContributor =
+			new BlogsAMEditorConfigContributor();
 
-		journalItemSelectorCriterion.setDesiredItemSelectorReturnTypes(
+		BlogsItemSelectorCriterion blogsItemSelectorCriterion =
+			new BlogsItemSelectorCriterion();
+
+		blogsItemSelectorCriterion.setDesiredItemSelectorReturnTypes(
 			new FileEntryItemSelectorReturnType());
 
-		amJournalEditorConfigContributor.
+		blogsAMEditorConfigContributor.
 			addAMImageFileEntryItemSelectorReturnType(
-				journalItemSelectorCriterion);
+				blogsItemSelectorCriterion);
 
 		List<ItemSelectorReturnType> desiredItemSelectorReturnTypes =
-			journalItemSelectorCriterion.getDesiredItemSelectorReturnTypes();
+			blogsItemSelectorCriterion.getDesiredItemSelectorReturnTypes();
 
 		Assert.assertEquals(
 			desiredItemSelectorReturnTypes.toString(), 2,
 			desiredItemSelectorReturnTypes.size());
-
-		_assertContains(
-			desiredItemSelectorReturnTypes,
-			itemSelectorReturnType -> itemSelectorReturnType instanceof
+		Assert.assertTrue(
+			desiredItemSelectorReturnTypes.get(0) instanceof
 				AMImageFileEntryItemSelectorReturnType);
+		Assert.assertTrue(
+			desiredItemSelectorReturnTypes.get(1) instanceof
+				FileEntryItemSelectorReturnType);
+	}
 
-		_assertContains(
-			desiredItemSelectorReturnTypes,
-			itemSelectorReturnType ->
-				itemSelectorReturnType instanceof
-					FileEntryItemSelectorReturnType);
+	@Test
+	public void testAMReturnTypeIsAddedToAllItemSelectorCriteria()
+		throws Exception {
+
+		ItemSelectorCriterion[] itemSelectorCriteria = {
+			_initializeItemSelectorCriterion(new BlogsItemSelectorCriterion()),
+			_initializeItemSelectorCriterion(new FileItemSelectorCriterion()),
+			_initializeItemSelectorCriterion(new ImageItemSelectorCriterion()),
+			_initializeItemSelectorCriterion(new UploadItemSelectorCriterion())
+		};
+
+		JSONObject jsonObject = JSONUtil.put(
+			"filebrowserImageBrowseLinkUrl", RandomTestUtil.randomString());
+
+		Mockito.when(
+			_itemSelector.getItemSelectorCriteria(Mockito.anyString())
+		).thenReturn(
+			Arrays.asList(itemSelectorCriteria)
+		);
+
+		Mockito.when(
+			_itemSelector.getItemSelectedEventName(Mockito.anyString())
+		).thenReturn(
+			RandomTestUtil.randomString()
+		);
+
+		Mockito.when(
+			_itemSelector.getItemSelectorURL(
+				Mockito.any(RequestBackedPortletURLFactory.class),
+				Mockito.anyString(), Mockito.<ItemSelectorCriterion>any())
+		).thenReturn(
+			_portletURL
+		);
+
+		Mockito.when(
+			_portletURL.toString()
+		).thenReturn(
+			RandomTestUtil.randomString()
+		);
+
+		BlogsAMEditorConfigContributor blogsAMEditorConfigContributor =
+			new BlogsAMEditorConfigContributor();
+
+		ReflectionTestUtil.setFieldValue(
+			blogsAMEditorConfigContributor, "_itemSelector", _itemSelector);
+
+		blogsAMEditorConfigContributor.populateConfigJSONObject(
+			jsonObject, _inputEditorTaglibAttributes, _themeDisplay,
+			_requestBackedPortletURLFactory);
+
+		for (ItemSelectorCriterion itemSelectorCriterion :
+				itemSelectorCriteria) {
+
+			for (ItemSelectorReturnType itemSelectorReturnType :
+					itemSelectorCriterion.getDesiredItemSelectorReturnTypes()) {
+
+				Assert.assertTrue(
+					itemSelectorReturnType instanceof
+						AMImageFileEntryItemSelectorReturnType);
+			}
+		}
 	}
 
 	@Test
@@ -260,10 +322,10 @@ public class AMJournalEditorConfigContributorTest {
 		JSONObject jsonObject = JSONFactoryUtil.createJSONObject(
 			originalJSONObject.toString());
 
-		AMJournalEditorConfigContributor amJournalEditorConfigContributor =
-			new AMJournalEditorConfigContributor();
+		BlogsAMEditorConfigContributor blogsAMEditorConfigContributor =
+			new BlogsAMEditorConfigContributor();
 
-		amJournalEditorConfigContributor.populateConfigJSONObject(
+		blogsAMEditorConfigContributor.populateConfigJSONObject(
 			jsonObject, _inputEditorTaglibAttributes, _themeDisplay,
 			_requestBackedPortletURLFactory);
 
@@ -281,10 +343,10 @@ public class AMJournalEditorConfigContributorTest {
 		JSONObject jsonObject = JSONFactoryUtil.createJSONObject(
 			originalJSONObject.toString());
 
-		AMJournalEditorConfigContributor amJournalEditorConfigContributor =
-			new AMJournalEditorConfigContributor();
+		BlogsAMEditorConfigContributor blogsAMEditorConfigContributor =
+			new BlogsAMEditorConfigContributor();
 
-		amJournalEditorConfigContributor.populateConfigJSONObject(
+		blogsAMEditorConfigContributor.populateConfigJSONObject(
 			jsonObject, _inputEditorTaglibAttributes, _themeDisplay,
 			_requestBackedPortletURLFactory);
 
@@ -295,22 +357,48 @@ public class AMJournalEditorConfigContributorTest {
 	}
 
 	@Test
-	public void testImgIsNotAddedToAllowedContentIfEverythingWasAlreadyAllowed()
+	public void testImgIsNotAddedToAllowedContentIfAlreadyPresent()
 		throws Exception {
 
-		JSONObject originalJSONObject = JSONUtil.put("allowedContent", true);
+		JSONObject originalJSONObject = JSONUtil.put(
+			"allowedContent", "a[*](*); div(*); img[*](*){*};");
 
 		JSONObject jsonObject = JSONFactoryUtil.createJSONObject(
 			originalJSONObject.toString());
 
-		AMJournalEditorConfigContributor amJournalEditorConfigContributor =
-			new AMJournalEditorConfigContributor();
+		BlogsAMEditorConfigContributor blogsAMEditorConfigContributor =
+			new BlogsAMEditorConfigContributor();
 
-		amJournalEditorConfigContributor.populateConfigJSONObject(
+		blogsAMEditorConfigContributor.populateConfigJSONObject(
 			jsonObject, _inputEditorTaglibAttributes, _themeDisplay,
 			_requestBackedPortletURLFactory);
 
-		JSONObject expectedJSONObject = JSONUtil.put("allowedContent", true);
+		JSONObject expectedJSONObject = JSONUtil.put(
+			"allowedContent", "a[*](*); div(*); img[*](*){*};");
+
+		JSONAssert.assertEquals(
+			expectedJSONObject.toString(), jsonObject.toString(), true);
+	}
+
+	@Test
+	public void testImgIsNotAddedToAllowedContentIfAnyContentAllowed()
+		throws Exception {
+
+		JSONObject originalJSONObject = JSONUtil.put(
+			"allowedContent", Boolean.TRUE.toString());
+
+		JSONObject jsonObject = JSONFactoryUtil.createJSONObject(
+			originalJSONObject.toString());
+
+		BlogsAMEditorConfigContributor blogsAMEditorConfigContributor =
+			new BlogsAMEditorConfigContributor();
+
+		blogsAMEditorConfigContributor.populateConfigJSONObject(
+			jsonObject, _inputEditorTaglibAttributes, _themeDisplay,
+			_requestBackedPortletURLFactory);
+
+		JSONObject expectedJSONObject = JSONUtil.put(
+			"allowedContent", Boolean.TRUE.toString());
 
 		JSONAssert.assertEquals(
 			expectedJSONObject.toString(), jsonObject.toString(), true);
@@ -326,13 +414,13 @@ public class AMJournalEditorConfigContributorTest {
 		JSONObject jsonObject = JSONFactoryUtil.createJSONObject(
 			originalJSONObject.toString());
 
-		AMJournalEditorConfigContributor amJournalEditorConfigContributor =
-			new AMJournalEditorConfigContributor();
+		BlogsAMEditorConfigContributor blogsAMEditorConfigContributor =
+			new BlogsAMEditorConfigContributor();
 
 		ReflectionTestUtil.setFieldValue(
-			amJournalEditorConfigContributor, "_itemSelector", _itemSelector);
+			blogsAMEditorConfigContributor, "_itemSelector", _itemSelector);
 
-		amJournalEditorConfigContributor.populateConfigJSONObject(
+		blogsAMEditorConfigContributor.populateConfigJSONObject(
 			jsonObject, _inputEditorTaglibAttributes, _themeDisplay,
 			_requestBackedPortletURLFactory);
 
@@ -368,13 +456,13 @@ public class AMJournalEditorConfigContributorTest {
 		JSONObject jsonObject = JSONFactoryUtil.createJSONObject(
 			originalJSONObject.toString());
 
-		AMJournalEditorConfigContributor amJournalEditorConfigContributor =
-			new AMJournalEditorConfigContributor();
+		BlogsAMEditorConfigContributor blogsAMEditorConfigContributor =
+			new BlogsAMEditorConfigContributor();
 
 		ReflectionTestUtil.setFieldValue(
-			amJournalEditorConfigContributor, "_itemSelector", _itemSelector);
+			blogsAMEditorConfigContributor, "_itemSelector", _itemSelector);
 
-		amJournalEditorConfigContributor.populateConfigJSONObject(
+		blogsAMEditorConfigContributor.populateConfigJSONObject(
 			jsonObject, _inputEditorTaglibAttributes, _themeDisplay,
 			_requestBackedPortletURLFactory);
 
@@ -420,25 +508,25 @@ public class AMJournalEditorConfigContributorTest {
 
 		Mockito.when(
 			_itemSelector.getItemSelectorCriteria(
-				"journalItemSelectorCriterionFileEntryItemSelectorReturnType")
+				"blogsItemSelectorCriterionFileEntryItemSelectorReturnType")
 		).thenReturn(
 			_getBlogsItemSelectorCriterionFileEntryItemSelectorReturnType()
 		);
 
 		JSONObject originalJSONObject = JSONUtil.put(
 			"filebrowserImageBrowseLinkUrl",
-			"journalItemSelectorCriterionFileEntryItemSelectorReturnType");
+			"blogsItemSelectorCriterionFileEntryItemSelectorReturnType");
 
 		JSONObject jsonObject = JSONFactoryUtil.createJSONObject(
 			originalJSONObject.toString());
 
-		AMJournalEditorConfigContributor amJournalEditorConfigContributor =
-			new AMJournalEditorConfigContributor();
+		BlogsAMEditorConfigContributor blogsAMEditorConfigContributor =
+			new BlogsAMEditorConfigContributor();
 
 		ReflectionTestUtil.setFieldValue(
-			amJournalEditorConfigContributor, "_itemSelector", _itemSelector);
+			blogsAMEditorConfigContributor, "_itemSelector", _itemSelector);
 
-		amJournalEditorConfigContributor.populateConfigJSONObject(
+		blogsAMEditorConfigContributor.populateConfigJSONObject(
 			jsonObject, _inputEditorTaglibAttributes, _themeDisplay,
 			_requestBackedPortletURLFactory);
 
@@ -490,13 +578,13 @@ public class AMJournalEditorConfigContributorTest {
 		JSONObject jsonObject = JSONFactoryUtil.createJSONObject(
 			originalJSONObject.toString());
 
-		AMJournalEditorConfigContributor amJournalEditorConfigContributor =
-			new AMJournalEditorConfigContributor();
+		BlogsAMEditorConfigContributor blogsAMEditorConfigContributor =
+			new BlogsAMEditorConfigContributor();
 
 		ReflectionTestUtil.setFieldValue(
-			amJournalEditorConfigContributor, "_itemSelector", _itemSelector);
+			blogsAMEditorConfigContributor, "_itemSelector", _itemSelector);
 
-		amJournalEditorConfigContributor.populateConfigJSONObject(
+		blogsAMEditorConfigContributor.populateConfigJSONObject(
 			jsonObject, _inputEditorTaglibAttributes, _themeDisplay,
 			_requestBackedPortletURLFactory);
 
@@ -508,24 +596,10 @@ public class AMJournalEditorConfigContributorTest {
 		);
 	}
 
-	private <T> void _assertContains(
-		Collection<T> collection, Predicate<T> predicate) {
-
-		boolean match = false;
-
-		for (T t : collection) {
-			if (predicate.test(t)) {
-				match = true;
-
-				break;
-			}
-		}
-
-		Assert.assertTrue(match);
-	}
-
 	private List<ItemSelectorCriterion>
 		_getAudioItemSelectorCriterionFileEntryItemSelectorReturnType() {
+
+		List<ItemSelectorCriterion> itemSelectorCriteria = new ArrayList<>();
 
 		AudioItemSelectorCriterion audioItemSelectorCriterion =
 			new AudioItemSelectorCriterion();
@@ -533,23 +607,31 @@ public class AMJournalEditorConfigContributorTest {
 		audioItemSelectorCriterion.setDesiredItemSelectorReturnTypes(
 			new FileEntryItemSelectorReturnType());
 
-		return Collections.singletonList(audioItemSelectorCriterion);
+		itemSelectorCriteria.add(audioItemSelectorCriterion);
+
+		return itemSelectorCriteria;
 	}
 
 	private List<ItemSelectorCriterion>
 		_getBlogsItemSelectorCriterionFileEntryItemSelectorReturnType() {
 
-		JournalItemSelectorCriterion journalItemSelectorCriterion =
-			new JournalItemSelectorCriterion();
+		List<ItemSelectorCriterion> itemSelectorCriteria = new ArrayList<>();
 
-		journalItemSelectorCriterion.setDesiredItemSelectorReturnTypes(
+		BlogsItemSelectorCriterion blogsItemSelectorCriterion =
+			new BlogsItemSelectorCriterion();
+
+		blogsItemSelectorCriterion.setDesiredItemSelectorReturnTypes(
 			new FileEntryItemSelectorReturnType());
 
-		return Collections.singletonList(journalItemSelectorCriterion);
+		itemSelectorCriteria.add(blogsItemSelectorCriterion);
+
+		return itemSelectorCriteria;
 	}
 
 	private List<ItemSelectorCriterion>
 		_getFileItemSelectorCriterionFileEntryItemSelectorReturnType() {
+
+		List<ItemSelectorCriterion> itemSelectorCriteria = new ArrayList<>();
 
 		FileItemSelectorCriterion fileItemSelectorCriterion =
 			new FileItemSelectorCriterion();
@@ -557,12 +639,24 @@ public class AMJournalEditorConfigContributorTest {
 		fileItemSelectorCriterion.setDesiredItemSelectorReturnTypes(
 			new FileEntryItemSelectorReturnType());
 
-		return Collections.singletonList(fileItemSelectorCriterion);
+		itemSelectorCriteria.add(fileItemSelectorCriterion);
+
+		return itemSelectorCriteria;
+	}
+
+	private ItemSelectorCriterion _initializeItemSelectorCriterion(
+		ItemSelectorCriterion itemSelectorCriterion) {
+
+		itemSelectorCriterion.setDesiredItemSelectorReturnTypes(
+			new ArrayList<ItemSelectorReturnType>());
+
+		return itemSelectorCriterion;
 	}
 
 	private final Map<String, Object> _inputEditorTaglibAttributes =
 		new HashMap<>();
 	private final ItemSelector _itemSelector = Mockito.mock(ItemSelector.class);
+	private final PortletURL _portletURL = Mockito.mock(PortletURL.class);
 	private final RequestBackedPortletURLFactory
 		_requestBackedPortletURLFactory = Mockito.mock(
 			RequestBackedPortletURLFactory.class);
